Reverted `fromListS` to use lists
The reasoning behind this is that if OverloadedLists becomes standard
then this would lead to an ambiguous type when passing in a list as an
argument.  It's better for the consumer to have a concrete input because
users can always use `Data.Foldable.toList` to convert to a list if
necessary.