Temporarily removed `WriterT` and `RWST` benchmarks
The change to `distribute` caused a performance regression singe it no longer
keeps the accumulator strictly evaluated.

I'm not sure whether I want to fix this or not.  Having a functional `writerP`
is not as urgent as it used to be now that one can combine multiple folds using
`foldl`.  Additionally, I would prefer the uniformity of using `distribute`
everywhere in `Pipes.Lift`.  I think the real solution is for `transformers` to
fix `WriterT`.