Factor out boilerplate in composition/instance functions
This makes the code easier to understand, since one can now more quickly ignore
the boilerplate without inspecting it "in case" it might be different. It also
makes the instance functions less dependent on some internal details of Proxy.

I marked all of the functions as INLINABLE so there should be no runtime cost.
(INLINE may be better, but I'm not familiar with Haskell internals to be able
to judge this by myself.)

Ideally we would write fixm,fixl,fixr as one function, but I couldn't figure
out how to satisfy the type checker in this way.