Switch `Pipes.Lift` to use `ExceptT` instead of `ErrorT`
Finally!  I've been waiting to make this change for a very long time.
`ExceptT` is sufficiently widespread and there has been a long enough
deprecation cycle to justify removing `ErrorT`.  This is a breaking change.

This will fix a lot of build warnings and also prepare `pipes` for the day
when `transformers` removes `ErrorT`.