Disable building benchmarks in Travis
`criterion` and `test-framework` do not mix well with each other on `ghc-7.6`.
The issue is that:

* `ghc-7.6` comes with `bytestring-0.10.0.2` installed by default
* `ghc-7.6` also comes with a `regex-posix` built against `bytestring-0.10.0.2`
* `criterion` requires a version of `bytestring` newer than `0.10.0.2` due to a
  dependency on `aeson`
* `test-framework` requires `regex-posix`

This leads to one of two problems when building both the benchmark suite and the
test suite at the same time for `ghc-7.6`.  Either:

* older versions of `cabal` fail at build time because they select the following
  invalid plan:
    * pick a `semigroups` library built against `bytestring-0.10.02` that is
      missing a `Semigroup` instance for `Data.ByteString.Builder.Builder`
      (this instance is only provided when building against newer versions of
       `bytestring`)
    * pick an `aeson` library built against `bytestring-0.10.8.1` that expects
      a `Semigroup` instance for `Data.ByteString.Builder.Builder`
* newer versions of `cabal` fail at plan time because they refuse to have two
  different versions of `bytestring` in the same build plan

My workaround for now is to disable building benchmarks in Travis.  I will
re-enable building benchmarks once I phase out support for `ghc-7.6.3` (which
in turn will happen when Debian Stable ships a newer `ghc` version)