Version 4.2.1 => 4.3.0.  Fixes #183
Older versions of `pipes-safe` also provides a `MonadThrow` and `MonadCatch`
instance for `Proxy`, but they have a `pipes < 4.3` bound.  In order to avoid
incompatibility between the two libraries the plan is to:

* Increase the `pipes` version to `4.3.0`
* Release a new `pipes-safe` version with a `pipes >= 4.3.0` constraint

This ensures that the two libraries will never have a build failure when built
together.  At worst they might incur a dependency resolution failure