Elaborate on the documentation for ListT
This implements several of the documentation ideas I brought up in #184.

The present problem, I think, is that it is not obvious how to accomplish
some basic ListT tasks (e.g. converting `ListT m a` to `m [a]`) unless
you have the habit of knowing to convert to Producer first.

I also wanted to emphasize some of the things that you can do using just
the typeclass instances.

I'm aiming to augment the ListT API documentation with just enough links
and suggestions to serve as jumping-off points, without cluttering up
the haddock page for the Pipes module too much.